complete new structure of game process
central element of game process ist class CivOptions. Each civilization generates an object. In this class every option of civilization in every game step and section is configured. 
All these options are distributed to other objects. 
On the other hand class GameStep (the central class before) is reduced to an minimum. It is degenerated common state class.