implementation of new class Drawable
Class Drawable will be the basis class of all Drawable Objects. 
All important properties of an Image are transposed here:
* image
* image constant
* image detail
* rotation
* cardNr
* map position
* visibility
* resize
* image Position
* scale

Furthermore, enum ImageConstant is implemented. This class should replace enum ImageObject. All image constants are set here as in ImageObject, but realizing is a little bit smarter

Military Unit is the first class with this implementation. It inherits of Drawable.